+++ src/main/java/com/Asis/api/domain/unidade/repository/UnidadeSUASRepository.java
@@ -3,5 +3,5 @@ package com.Asis.api.domain.unidade.repository;
 import com.Asis.api.domain.unidade.entity.UnidadeSUASEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UnidadeSUASRepository extends JpaRepository<UnidadeSUASEntity, Long> {
+public interface UnidadeSUASRepository extends JpaRepository<UnidadeSUASEntity, Integer> {
 }
