+++ src/main/resources/db/migration/V9__create_table_unidade_funcionario.sql
@@ -1,7 +1,7 @@
 -- Criando tabela intermédiaria entre unidade e funcionário
 
 CREATE TABLE tb_unidade_funcionario(
-    unidade_id BIGINT NOT NULL,
+    unidade_id INTEGER NOT NULL,
     usuario_id VARCHAR(36) NOT NULL,
     funcao VARCHAR(45),
     PRIMARY KEY(unidade_id, usuario_id),
